src/requirements.txt
moment
munch
google-cloud-pubsub
mementos
python-dotenv